% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_menu.R
\name{check_exit}
\alias{check_exit}
\title{Verifica si el usuario quiere salir}
\usage{
check_exit(input)
}
\arguments{
\item{input}{String ingresado por el usuario.}
}
\value{
TRUE si el usuario escribió 'exit', FALSE en caso contrario.
}
\description{
Verifica si el usuario quiere salir
}
