% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_render.R
\name{.add_missing_to_legend_data}
\alias{.add_missing_to_legend_data}
\title{Appends the missing / inapplicable entries to grouped legend data}
\usage{
.add_missing_to_legend_data(ley_data, config)
}
\arguments{
\item{ley_data}{Output of \code{.build_legend_data()}.}

\item{config}{Configuration list.}
}
\value{
\code{ley_data} with the extra entries.
}
\description{
Appends the missing / inapplicable entries to grouped legend data
}
\keyword{internal}
