% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{.detect_matrix_format}
\alias{.detect_matrix_format}
\title{Detects the format of a character matrix file}
\usage{
.detect_matrix_format(path)
}
\arguments{
\item{path}{Path to the file.}
}
\value{
"tnt", "nexus" or "csv".
}
\description{
Detects the format of a character matrix file
}
