% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_multimapr_tree.R
\name{.emtree_abrir_dispositivo}
\alias{.emtree_abrir_dispositivo}
\title{Abre el dispositivo gráfico adecuado según \code{format}}
\usage{
.emtree_abrir_dispositivo(filename, format, ancho_in, alto_in)
}
\value{
Invisible NULL.  El dispositivo queda abierto tras la llamada.
}
\description{
Abre el dispositivo gráfico adecuado según \code{format}
}
