% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_multimapr_tree.R
\name{.emtree_calc_offsets}
\alias{.emtree_calc_offsets}
\title{Computes the offset vector from the current physical device parameters
(par("usr") / par("pin")).}
\usage{
.emtree_calc_offsets(N, offset_range = 0.1)
}
\arguments{
\item{N}{Number of histories.}

\item{offset_range}{Amplitude of the Y range (default 0.1).}
}
\value{
List with \code{dx} and \code{dy} (vectors of length N).
}
\description{
Computes the offset vector from the current physical device parameters
(par("usr") / par("pin")).
}
