% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_multimapr_tree.R
\name{.emtree_calcular_desfases}
\alias{.emtree_calcular_desfases}
\title{Calcula el vector de desfases a partir de los parámetros físicos actuales
del dispositivo (par("usr") / par("pin")).}
\usage{
.emtree_calcular_desfases(N, rango_desfase = 0.1)
}
\arguments{
\item{N}{Número de historias.}

\item{rango_desfase}{Amplitud del rango en Y (default 0.1).}
}
\value{
Lista con \code{dx} y \code{dy} (vectores de longitud N).
}
\description{
Calcula el vector de desfases a partir de los parámetros físicos actuales
del dispositivo (par("usr") / par("pin")).
}
