% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_multimapr_tree.R
\name{.emtree_get_PlotPhyloEnv}
\alias{.emtree_get_PlotPhyloEnv}
\title{Recupera .PlotPhyloEnv desde el namespace de ape de forma segura}
\usage{
.emtree_get_PlotPhyloEnv()
}
\description{
Recupera .PlotPhyloEnv desde el namespace de ape de forma segura
}
