% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_multimapr_tree.R
\name{.emtree_validar_arbol}
\alias{.emtree_validar_arbol}
\title{Verifica que \code{tree} sea un objeto phylo válido con aristas y tip.labels}
\usage{
.emtree_validar_arbol(tree)
}
\description{
Verifica que \code{tree} sea un objeto phylo válido con aristas y tip.labels
}
