% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_multimapr_tree.R
\name{.emtree_validar_color_list}
\alias{.emtree_validar_color_list}
\title{Verifica que \code{color_list} sea coherente con el número de aristas del árbol}
\usage{
.emtree_validar_color_list(color_list, n_edges)
}
\description{
Verifica que \code{color_list} sea coherente con el número de aristas del árbol
}
