% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_multimapr_tree.R
\name{.emtree_validar_filename}
\alias{.emtree_validar_filename}
\title{Verifica que \code{filename} sea un string no vacío (sin extensión requerida)}
\usage{
.emtree_validar_filename(filename)
}
\description{
Verifica que \code{filename} sea un string no vacío (sin extensión requerida)
}
