% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_multimapr_tree.R
\name{.emtree_validar_tipo_formato}
\alias{.emtree_validar_tipo_formato}
\title{Verifica argumentos escalares de tipo y formato}
\usage{
.emtree_validar_tipo_formato(type, format)
}
\description{
Verifica argumentos escalares de tipo y formato
}
