% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_multimapr_tree.R
\name{.emtree_validate_node_size}
\alias{.emtree_validate_node_size}
\title{Checks the internal-node figure size multiplier}
\usage{
.emtree_validate_node_size(node_size)
}
\arguments{
\item{node_size}{Positive number.}
}
\description{
Checks the internal-node figure size multiplier
}
\keyword{internal}
