% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_multimapr_tree.R
\name{.emtree_validate_tree}
\alias{.emtree_validate_tree}
\title{Checks that \code{tree} is a valid phylo object with edges and tip.labels}
\usage{
.emtree_validate_tree(tree)
}
\description{
Checks that \code{tree} is a valid phylo object with edges and tip.labels
}
