% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_multimapr_tree.R
\name{.emtree_validate_type_format}
\alias{.emtree_validate_type_format}
\title{Checks scalar type and format arguments}
\usage{
.emtree_validate_type_format(type, format)
}
\description{
Checks scalar type and format arguments
}
