% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiMapR.R
\name{.exportar_dispositivo}
\alias{.exportar_dispositivo}
\title{Orquestador principal de MultiMapR}
\usage{
.exportar_dispositivo(
  filename,
  format,
  filogenia,
  tipo_arbol,
  expr,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{filogenia}{Ruta al archivo de árbol o un objeto phylo.}

\item{datos_caracteres}{Ruta al archivo CSV o un data.frame.}

\item{grosor}{Grosor opcional de las ramas.}

\item{label_offset}{Desfase dinámico de las etiquetas.}
}
\description{
Orquestador principal de MultiMapR
}
