% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_alg.R
\name{.get_descendants}
\alias{.get_descendants}
\title{Gets all descendants of a node (iterative)}
\usage{
.get_descendants(tree, node)
}
\arguments{
\item{tree}{phylo object.}

\item{node}{Index of the starting node.}
}
\value{
Integer vector of descendant node indices.
}
\description{
Iterative version of getDescendants() that avoids recursion problems
in large trees and does not use <<-.
}
