% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_multimapr_tree.R
\name{.legend_title_height}
\alias{.legend_title_height}
\title{Height reserved for a legend block title (data units, current device)}
\usage{
.legend_title_height(cex)
}
\arguments{
\item{cex}{Legend text size.}
}
\description{
Height reserved for a legend block title (data units, current device)
}
\keyword{internal}
