% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{.quitar_bom}
\alias{.quitar_bom}
\title{Elimina el BOM UTF-8 (EF BB BF) si existe y devuelve la ruta a usar}
\usage{
.quitar_bom(ruta_csv)
}
\arguments{
\item{ruta_csv}{Ruta al archivo original.}
}
\value{
Ruta al archivo limpio (temporal o la original).
}
\description{
Si el archivo tiene BOM, copia el contenido sin los 3 bytes iniciales a un
archivo temporal y devuelve su ruta. Si no tiene BOM devuelve la ruta original.
El archivo temporal se elimina automáticamente al terminar la sesión de R.
}
