% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{.strip_quotes}
\alias{.strip_quotes}
\title{Removes surrounding single or double quotes (and whitespace) from names}
\usage{
.strip_quotes(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector.
}
\description{
Removes surrounding single or double quotes (and whitespace) from names
}
