% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Operador null-coalescing: devuelve x si no es NULL, y en caso contrario}
\usage{
x \%||\% y
}
\description{
Operador null-coalescing: devuelve x si no es NULL, y en caso contrario
}
