% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_render.R
\name{init_edge_colors}
\alias{init_edge_colors}
\title{Initializes the edge color vector in gray}
\usage{
init_edge_colors(filogenia)
}
\arguments{
\item{filogenia}{phylo object.}
}
\value{
Named character vector of length nrow(filogenia$edge).
}
\description{
Initializes the edge color vector in gray
}
