% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{op-null-default}
\alias{op-null-default}
\alias{\%||\%}
\title{Null-coalescing operator: returns x if not NULL, otherwise y}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Value to test; returned as-is if not \code{NULL}.}

\item{y}{Fallback value returned when \code{x} is \code{NULL}.}
}
\value{
\code{x} if \code{!is.null(x)}, otherwise \code{y}.
}
\description{
Null-coalescing operator: returns x if not NULL, otherwise y
}
\keyword{internal}
