% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_render.R
\name{plot_fan_capa}
\alias{plot_fan_capa}
\title{Replots the fan tree with slight rotation and displacement for additional layers}
\usage{
plot_fan_capa(
  filogenia,
  edge_colors,
  grosor,
  alpha,
  escala,
  rotacion,
  offset_x = 0,
  offset_y = 0,
  cex_aj = 0.5,
  escala_relativa = 1
)
}
\description{
Replots the fan tree with slight rotation and displacement for additional layers
}
