% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_menu.R
\name{check_exit}
\alias{check_exit}
\title{Checks whether the user wants to exit}
\usage{
check_exit(input)
}
\arguments{
\item{input}{String entered by the user.}
}
\value{
TRUE if the user typed 'exit', FALSE otherwise.
}
\description{
Checks whether the user wants to exit
}
