% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_multimapr_tree.R
\name{.emtree_get_PlotPhyloEnv}
\alias{.emtree_get_PlotPhyloEnv}
\title{Safely retrieves .PlotPhyloEnv from the ape namespace}
\usage{
.emtree_get_PlotPhyloEnv()
}
\description{
Safely retrieves .PlotPhyloEnv from the ape namespace
}
