% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_multimapr_tree.R
\name{.emtree_open_device}
\alias{.emtree_open_device}
\title{Opens the appropriate graphics device according to \code{format}}
\usage{
.emtree_open_device(filename, format, width_in, height_in)
}
\value{
Invisible NULL. The device remains open after the call.
}
\description{
Opens the appropriate graphics device according to \code{format}
}
