% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_multimapr_tree.R
\name{.emtree_validate_color_list}
\alias{.emtree_validate_color_list}
\title{Checks that \code{color_list} is consistent with the number of tree edges}
\usage{
.emtree_validate_color_list(color_list, n_edges)
}
\description{
Checks that \code{color_list} is consistent with the number of tree edges
}
