% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiMapR.R
\name{.export_device}
\alias{.export_device}
\title{Main orchestrator of MultiMapR}
\usage{
.export_device(
  filename,
  format,
  phylogeny,
  tree_type,
  expr,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{phylogeny}{Path to the tree file or a phylo object.}

\item{character_data}{Path to the CSV file or a data.frame.}

\item{branch_width}{Optional branch width.}

\item{label_offset}{Dynamic label offset.}
}
\description{
Main orchestrator of MultiMapR
}
