% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{.normalize_char_names}
\alias{.normalize_char_names}
\title{Generates R-safe character names}
\usage{
.normalize_char_names(raw_names, n)
}
\arguments{
\item{raw_names}{Vector with the raw names of the character columns.
NULL or empty -> automatic names are generated.}

\item{n}{Number of character columns.}
}
\value{
Character vector of length n.
}
\description{
If the CSV has a header, those names are used (sanitized).
If there is no header, generates "char1", "char2", ...
}
