% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{.normalize_polymorphism}
\alias{.normalize_polymorphism}
\title{Converts polymorphism notations to the "0/1" form used by Fitch}
\usage{
.normalize_polymorphism(x)
}
\arguments{
\item{x}{Character vector of states.}
}
\value{
Character vector of the same length.
}
\description{
\verb{[01]}, \code{{01}}, \code{(01)}, \verb{[0 1]}, \verb{(0,1)}, \code{{0/1}}  ->  \code{0/1}
\verb{[0]}                                              ->  \code{0}
Any other value is returned unchanged (trimmed).
}
