% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{.sanitize_tree}
\alias{.sanitize_tree}
\title{Sanitizes a tree file for compatibility with ape}
\usage{
.sanitize_tree(tree_path)
}
\arguments{
\item{tree_path}{Path to the original file.}
}
\value{
Path to the clean file (temporary if changes were made, original otherwise).
}
\description{
Removes common problems produced by different phylogenetic programs:
\itemize{
\item Windows line endings (CR+LF converted to LF)
\item Numbered comments in Mesquite TRANSLATE blocks (\verb{[0]}, \verb{[1]}, ...)
that ape cannot parse
}
}
