% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_render.R
\name{plot_cladogram_split}
\alias{plot_cladogram_split}
\title{Draws an offset cladogram for the second character onward}
\usage{
plot_cladogram_split(
  tree,
  base_coords,
  edge_colors,
  grosor2,
  alpha2,
  offset_clado
)
}
\description{
Draws an offset cladogram for the second character onward
}
