% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_menu.R
\name{prompt_character_colors}
\alias{prompt_character_colors}
\title{Prompts the user for colors for all states of a character}
\usage{
prompt_character_colors(sorted_states, character_name)
}
\arguments{
\item{sorted_states}{Vector of already-sorted states.}

\item{character_name}{String with the character name (for the header).}
}
\value{
Named character vector state -> color.
}
\description{
Iterates over each state in \code{sorted_states} calling
\code{prompt_state_color()} and builds a named vector state -> color.
}
