% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{read_csv_characters}
\alias{read_csv_characters}
\title{Reads a character CSV file and prepares it for MultiMapR}
\usage{
read_csv_characters(
  ruta_csv,
  sep = ",",
  col_especies = 1,
  normalizar_espacios = FALSE
)
}
\arguments{
\item{ruta_csv}{Ruta al archivo CSV.}

\item{sep}{Separador de campos (default ",").}

\item{col_especies}{Índice o nombre de la columna de especies (default 1).}

\item{normalizar_espacios}{Si TRUE, reemplaza espacios por "_" en los nombres
de especies (default FALSE).}
}
\value{
Data.frame con columna "Species" y columnas "charN" o nombres propios.
}
\description{
Reads a character CSV file and prepares it for MultiMapR
}
