% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{read_tree}
\alias{read_tree}
\title{Reads a phylogenetic tree file (Newick or NEXUS)}
\usage{
read_tree(tree_path, tree_format = "auto")
}
\arguments{
\item{tree_path}{Path to the tree file.}

\item{tree_format}{"auto" (default), "newick" or "nexus".}
}
\value{
phylo object.
}
\description{
Compatible with files from Mesquite, TNT, WinClada and other programs
that produce non-standard variants of the NEXUS/Newick format.
}
